+++ packages/yoteams-build-core/CHANGELOG.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [*Unreleased*]- <*2021-06-10*>
+## [*Unreleased*]- <*2021-09-13*>
+
+### Added
+
+* Added flag `--no-schema-validation` to skip validation of Manifest schema (not listed as a supported schema)
 
 ### Fixes